Restructure existing plugins
