+++ src/djpress/plugins/hook_registry.py
@@ -1,0 +1,15 @@
+"""Plugin system for DJ Press."""
+
+import logging
+from enum import Enum
+
+logger = logging.getLogger(__name__)
+
+
+# Hook definitions
+class Hooks(Enum):
+    """Available hook points in DJ Press."""
+
+    PRE_RENDER_CONTENT = "pre_render_content"
+    POST_RENDER_CONTENT = "post_render_content"
+    POST_SAVE_POST = "post_save_post"
